+++ packages/babel-preset-veui/test/fixtures/source.js
@@ -1,0 +1,13 @@
+import { Input } from 'veui'
+
+export let a = {...{ a:  1 }}
+
+export class A {
+  a = 42
+}
+
+export default {
+  render () {
+    return <div>A</div>
+  }
+}